retail_ai_location_strategy_adk/requirements.txt
# Retail AI Location Strategy - ADK Dependencies
google-adk>=1.20.0
google-genai>=1.53.0
googlemaps>=4.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
